+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.0.3 - 2024-04-15
+
+### Changed
+- Correctly turned on `StrictConcurrency` in Swift 5.10 and earlier and adeed non-breaking conformance to `Sendable`.
+- Updated documentation for closure based usage where `nonisolated(unsafe)` is required to avoid an error in projects that use `StrictConcurrency`.
+
 ## 3.0.2 - 2024-02-07
 
 ### Added